refactor and simplify getting comment counts for articles
