+++ pages/index.vue
@@ -1,7 +1,6 @@
 <script setup lang="ts">
 import VCard from '@nypublicradio/nypr-design-system-vue3/v2/src/components/VCard.vue'
-import { useFetchCommentCounts } from '~~/composables/comments';
-import { useCommentCounts } from '~~/composables/states';
+import { useUpdateCommentCounts } from '~~/composables/comments';
 import useImageUrl from '~~/composables/useImageUrl'
 
 const articlesPromise = findArticlePages({}).then(({ data }) =>
@@ -44,16 +43,11 @@ const newsletterSubmitEvent = () => {
 }
 const navigation = useNavigation()
 
-onMounted(async () => {
+onMounted(() => {
   $analytics.sendPageView({ page_type: 'home_page' })
-  const commentCounts = useCommentCounts()
   const collectionArticles = homePageCollections.reduce((pages, collection) => [...pages, ...collection.data],[])
   const allArticles = [...articles, ...collectionArticles]
-  const commentIds = allArticles.map(article => String(article.legacyId || article.uuid))
-  const commentCountData = await useFetchCommentCounts(commentIds)
-  Object.entries(commentCountData).forEach(([key, value]) => {
-    commentCounts.value[key] =  value
-  })
+  useUpdateCommentCounts(allArticles)
 })
 
 const loadedNativoElements = []
